.bazaar/Sound: Fixes for MSVC
git-svn-id: svn://ultimatepp.org/upp/trunk@3455 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ bazaar/SoundExample/Wave.h
@@ -67,7 +67,7 @@ private:
 		f1=~freq1;f2=~freq2;
 		a1=~amp1; a2=~amp2;
 		sr=~samplerate;
-		snd.SetSampleRate(sr).SetFramesPerBuffer((fpb>0)?fpb:1);
+		snd.SetSampleRate(sr).SetFramesPerBuffer((fpb>0)?(int)fpb:1);
 		snd.Open(Null,StreamParameters(~device,2,SND_FLOAT32,SoundDevice(~device).LowOutputLatency));
 		if(CheckError()) return;
 		snd.Start();
